Reformat and move framework code into core
