+++ crates/enclave/core/src/chain_client.rs
@@ -1,3 +1,5 @@
+pub mod default;
+
 pub trait ChainClient: Send + Sync + 'static {
     const CHAIN_ID: &'static str;
