+++ crates/enclave/core/src/key_manager.rs
@@ -1,3 +1,6 @@
+pub mod default;
+pub mod shared;
+
 #[async_trait::async_trait]
 pub trait KeyManager: Send + Sync + 'static {
     type PubKey;
