+++ crates/enclave/core/src/key_manager/shared.rs
@@ -1,0 +1,31 @@
+use std::sync::Arc;
+
+use tokio::sync::RwLock;
+
+use crate::key_manager::KeyManager;
+
+#[derive(Clone, Debug)]
+pub struct SharedKeyManager<K> {
+    inner: Arc<RwLock<K>>,
+}
+
+impl<K> SharedKeyManager<K> {
+    pub fn wrapping(key_manager: K) -> Self {
+        Self {
+            inner: Arc::new(RwLock::new(key_manager)),
+        }
+    }
+}
+
+#[async_trait::async_trait]
+impl<K: KeyManager> KeyManager for SharedKeyManager<K> {
+    type PubKey = K::PubKey;
+
+    async fn keygen(&mut self) {
+        self.inner.write().await.keygen().await
+    }
+
+    async fn pub_key(&self) -> Option<Self::PubKey> {
+        self.inner.read().await.pub_key().await
+    }
+}
