+++ crates/enclave/core/src/kv_store.rs
@@ -5,6 +5,11 @@ use std::{
 
 use quartz_contract_core::state::{Config, Nonce};
 
+pub mod bincode;
+pub mod default;
+pub mod mpsc;
+pub mod shared;
+
 #[async_trait::async_trait]
 pub trait KvStore<K, V>: Send + Sync + 'static {
     type Error: ToString;
